restricts contacts' accessing by non-owner

+++ services/contacts.js
@@ -1,12 +1,9 @@
 const { Contact } = require('../model');
 
-const listContacts = async ({ limit = 5, page = 1, favorite = false }) => {
-  const { docs, totalDocs } = await Contact.paginate({}, {
+const listContacts = async (userId, { limit = 5, page = 1, favorite = false }) => {
+  const { docs, totalDocs } = await Contact.paginate({ owner: userId }, {
     limit,
     page,
-    populate: {
-      path: 'owner',
-    }
   });
 
   const contacts = favorite ? docs.filter(doc => doc.favorite) : docs;
@@ -15,28 +12,28 @@ const listContacts = async ({ limit = 5, page = 1, favorite = false }) => {
   return { contacts, total, limit: Number(limit), page: Number(page) };
 };
 
-const getContactById = async (id) => {
-  const result = await Contact.findById({ _id: id });
+const getContactById = async (userId, id) => {
+  const result = await Contact.findById({ _id: id, owner: userId });
   return result;
 };
 
-const createContact = async (contact, userId) => {
-  const result = await Contact.create({...contact, owner: userId});
+const createContact = async (userId, contact) => {
+  const result = await Contact.create({ ...contact, owner: userId });
   return result;
 };
 
-const updateContact = async (id, fields) => {
-  const result = await Contact.findByIdAndUpdate({ _id: id }, fields, { new: true });
+const updateContact = async (userId, id, fields) => {
+  const result = await Contact.findByIdAndUpdate({ _id: id, owner: userId }, fields, { new: true });
   return result;
 };
 
-const removeContact = async (id) => {
-  const result = await Contact.findByIdAndRemove({ _id: id });
+const removeContact = async (userId, id) => {
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: userId });
   return result;
 };
 
-const updateFavorite = async (id, { favorite }) => {
-  const result = Contact.findByIdAndUpdate({ _id: id }, { favorite }, { new: true });
+const updateFavorite = async (userId, id, { favorite }) => {
+  const result = Contact.findByIdAndUpdate({ _id: id, owner: userId }, { favorite }, { new: true });
   return result;
 };
